+++ src/section.h
@@ -8,9 +8,9 @@ public:
     enum {
         Type = QTreeWidgetItem::UserType + 1
     };
-    Section(QTreeWidget* parent = 0, quint8 id = 0, quint32 size = 0, quint32 offset = 0);
-    quint8 getId() const;
-    void setId(const quint8& value);
+    Section(QTreeWidget* parent = 0, quint32 id = 0, quint32 size = 0, quint32 offset = 0);
+    quint32 getId() const;
+    void setId(const quint32& value);
 
     quint32 getSize() const;
     void setSize(const quint32& value);
@@ -26,7 +26,7 @@ public:
     ~Section();
 
 private:
-    quint8 id;
+    quint32 id;
     quint32 size;
     quint32 offset;
 };
